requirements.txt
python-dotenv==0.21.0
SQLAlchemy==1.4.41
slack-bolt==1.16.2
slack-sdk==3.20.0
APScheduler==3.9.1.post1